convert types

+++ src/main/java/btt_telecom/api/modules/funcionario/controller/FuncionarioController.java
@@ -54,7 +54,7 @@ public class FuncionarioController extends AbstractMethods{
 	
 	
 	@PostMapping(path = "/page")
-	private ResponseEntity<Map<String, Object>> searchWithPage(@RequestParam(name = "value", defaultValue = "") String value, @RequestParam(name = "size") int size, @RequestParam(name = "page") int page) throws SQLException{
+	private ResponseEntity<Map<String, Object>> searchWithPage(@RequestParam(name = "value", defaultValue = "") String value, @RequestParam(name = "size") Long size, @RequestParam(name = "page") Long page) throws SQLException{
 		try {
 			if(value.equals("")) {
 				return new ResponseEntity<>(convertListToPage(funcionarioDAO.findAll(), size, page), HttpStatus.OK);
